# openapi-appengine.yaml
swagger: '2.0'
info:
  description: This is a simple API
  version: 1.0.0
  title: Spin 4 Athletics API
  # put the contact info for your development or API team
  contact:
    email: m.gammaldi@gmail.com

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

# tags are used for organizing operations
...
paths:
  /admin/sessions:
    post:
      tags:
      - admins
      operationId: addSessions
      summary: upload sessions 
      description: upload one or many training sessions 
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: sessionList
        description: session items to add
        schema:
            type: object
            properties:
              payload:
                type: array
                maxItems: 1000
                minItems: 1
                items:
                  $ref: "#/definitions/SessionItem"                 
      responses:
        201:
          description: session(s) created
        400:
          description: invalid input, object invalid
        403:
          description: forbidden
      security:
        - ApiKeyAuth: []  
    get:
      tags:
      - admins
      operationId: getSessions
      summary: searches sessions 
      description: |
        By passing in the appropriate options, you can search for
        available sessions in the system 
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: query
        name: searchString
        description: pass an optional search string for looking up session
        required: false
        type: string
      - in: query
        name: skip
        description: number of records to skip for pagination
        type: integer
        format: int32
        minimum: 0
      - in: query
        name: limit
        description: maximum number of records to return
        type: integer
        format: int32
        minimum: 0
        maximum: 50
      responses:
        200:
          description: search results matching criteria
          schema:
            type: array
            items:
              $ref: '#/definitions/SessionItem'
        400:
          description: bad input parameter
      security:
        - ApiKeyAuth: []
...
definitions:
  SessionItem:
    type: object
    required:
    - Name
    properties:
      Name: 
        type: string
        example : DI LIVIO
...
securityDefinitions:
  ApiKeyAuth:
    description: security applied to developers 
    name: key
    in: query
    type: apiKey
  OAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://example.com/oauth/authorize
    tokenUrl: https://example.com/oauth/token
    scopes:
      read: Grants read access
      write: Grants write access
      admin: Grants read and write access to administrative information
# Added by API Auto Mocking Plugin
# host: spin4athletics.appspot.com
schemes:
 - https
# Added by API Auto Mocking Plugin
# basePath: /v1
# Added by API Auto Mocking Plugin
host: spin4athletics.appspot.com
basePath: /v1